fix: i created a fresh folder for my own page. fix:  i removed the duplicates in the mockdata
