--- src/pages/JobProfile/components/mockData2.js
@@ -1,65 +1,0 @@
-const jobSkills = [
-    'JavaScript',
-    'React.js',
-    'Node.js',
-    'HTML5',
-    'CSS3',
-    'Python',
-    'Java',
-    'C#',
-    'Angular',
-    'Vue.js',
-    'SQL',
-    'MongoDB',
-    'Git',
-    'Redux',
-    'RESTful API',
-    'Responsive Design',
-    'Agile Development',
-    'Problem Solving',
-    'HTML5',
-    'CSS3',
-    'Python',
-    'Java',
-    'C#',
-    'Angular',
-    'Vue.js',
-    'SQL',
-    'MongoDB',
-    'Git',
-    'Redux',
-    'RESTful API',
-    'Python',
-    'Java',
-    'C#',
-    'Angular',
-    'Vue.js',
-    'SQL',
-    'MongoDB',
-    'Git',
-    'Redux',
-    'RESTful API',
-    'Python',
-    'Java',
-    'C#',
-    'Angular',
-    'Vue.js',
-    'SQL',
-    'MongoDB',
-    'Git',
-    'Redux',
-    'RESTful API',
-    'RESTful API',
-    'Python',
-    'Java',
-    'C#',
-    'Angular',
-    'Vue.js',
-    'SQL',
-    'MongoDB',
-    'Git',
-    'Redux',
-    'RESTful API',
-  ];
-
-  export default jobSkills;
