+++ src/pages/Skills/mockData.js
@@ -1,0 +1,22 @@
+const jobSkills = [
+    'JavaScript',
+    'React.js',
+    'Node.js',
+    'HTML5',
+    'CSS3',
+    'Python',
+    'Java',
+    'C#',
+    'Angular',
+    'Vue.js',
+    'SQL',
+    'MongoDB',
+    'Git',
+    'Redux',
+    'RESTful API',
+    'Responsive Design',
+    'Agile Development',
+    'Problem Solving',
+  ];
+
+  export default jobSkills;
