+++ src/routes.jsx
@@ -9,7 +9,7 @@ import JobProfilePage from "./pages/JobProfile/JobProfilePage";
 import LinkedInLearning from "./pages/LinkedInLearning/LinkedInLearningPage";
 import LoginPage from "./pages/Login/LoginPage";
 import TrainingPlanPage from "./pages/TrainingPlan/TrainingPlanPage";
-import SkillSetPage from "./pages/JobProfile/SkillSetPage";
+import SkillSetPage from "./pages/Skills/SkillSetPage";
 
 
 const ROUTES = [
